add .npmignore and move ts interfaces to separate file
