+++ .npmignore
@@ -1,0 +1,17 @@
+src
+public
+static
+.vscode
+.github
+node_modules
+.babelrc
+.eslintignore
+.eslintrc.js
+.gitignore
+.npmignore
+.travis.yml
+CONTRIBUTING.md
+jest.config.js
+package-lock.json
+README.md
+rollup.config.js
