+++ interfaces.d.ts
@@ -1,0 +1,40 @@
+export declare interface IField {
+    name: string;
+    caption: string;
+    operators: any;
+  }
+  
+  export declare interface IGroup {
+    name: string;
+    caption: string;
+  }
+  
+  export declare interface IOperator {
+    name: string;
+    caption: string;
+  }
+  
+  export declare interface IFilterValueItem {
+    key: any;
+    field: string;
+    operator: string;
+    value: any;
+  }
+  
+  export declare interface IFilterControlFilterValueGroup {
+    key: any;
+    groupName: string;
+    items: Array<IFilterControlFilterValue | IFilterValueItem>;
+  }
+  
+  export declare interface IFilterControlFilterValue {
+    groupName: string;
+    items: Array<IFilterControlFilterValue | IFilterValueItem>;
+  }
+  
+  export declare interface IFilterControlProps {
+    fields: IField[];
+    groups?: IGroup[];
+    filterValue: IFilterControlFilterValue;
+    onFilterValueChanged: (value: any) => void;
+  }
